Remove Capital namespace
